fix: rename file suffixes
