--- GestureHandlerButton.native.js
@@ -1,8 +1,0 @@
-import { requireNativeComponent } from 'react-native';
-
-const RNGestureHandlerButton = requireNativeComponent(
-  'RNGestureHandlerButton',
-  null
-);
-
-export default RNGestureHandlerButton;
